Добавляет ДЗ 5-1
