+++ js/util.js
@@ -1,0 +1,12 @@
+const getRandomNumber = (min, max) => {
+  if (min < 0 || max < 0 || min === max) {
+    return NaN;
+  }
+
+  if (min > max) {
+    [min, max] = [max, min];
+  }
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+export {getRandomNumber};
